@@ -65,27 +65,29 @@
     "url": "https://github.com/ng-bootstrap/ng-bootstrap/issues"
   },
   "homepage": "https://github.com/ng-bootstrap/ng-bootstrap#readme",
-  "dependencies": {},
+  "dependencies": {
+    "tslib": "^2.0.0"
+  },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.1",
-    "@angular-devkit/build-ng-packagr": "~0.900.1",
-    "@angular-devkit/core": "~9.0.0",
-    "@angular-devkit/schematics": "~9.0.0",
-    "@angular/animations": "~9.0.0",
-    "@angular/cli": "~9.0.1",
-    "@angular/common": "~9.0.0",
-    "@angular/compiler": "~9.0.0",
-    "@angular/compiler-cli": "~9.0.0",
-    "@angular/core": "~9.0.0",
-    "@angular/forms": "~9.0.0",
-    "@angular/localize": "~9.0.0",
-    "@angular/platform-browser": "~9.0.0",
-    "@angular/platform-browser-dynamic": "~9.0.0",
-    "@angular/platform-server": "~9.0.0",
-    "@angular/router": "~9.0.0",
+    "@angular-devkit/build-angular": "~0.1000.0",
+    "@angular-devkit/build-ng-packagr": "~0.1000.0",
+    "@angular-devkit/core": "~10.0.0",
+    "@angular-devkit/schematics": "~10.0.0",
+    "@angular/animations": "~10.0.0",
+    "@angular/cli": "~10.0.0",
+    "@angular/common": "~10.0.0",
+    "@angular/compiler": "~10.0.0",
+    "@angular/compiler-cli": "~10.0.0",
+    "@angular/core": "~10.0.0",
+    "@angular/forms": "~10.0.0",
+    "@angular/localize": "~10.0.0",
+    "@angular/platform-browser": "~10.0.0",
+    "@angular/platform-browser-dynamic": "~10.0.0",
+    "@angular/platform-server": "~10.0.0",
+    "@angular/router": "~10.0.0",
     "@commitlint/cli": "8.3.4",
     "@commitlint/config-angular": "8.3.4",
-    "@nguniversal/express-engine": "~9.0.0",
+    "@nguniversal/express-engine": "~10.0.0",
     "@schematics/angular": "~9.0.0",
     "@types/express": "^4.16.1",
     "@types/fs-extra": "^7.0.0",
@@ -102,6 +104,7 @@
     "conventional-changelog-cli": "^2.0.12",
     "core-js": "^2",
     "cross-spawn": "^7.0.1",
+    "coverage-istanbul-loader": "^3.0.3",
     "ejs": "2.6.1",
     "express": "^4.16.4",
     "fs-extra": "^8.0.0",
@@ -110,30 +113,29 @@
     "husky": "4.0.7",
     "jasmine": "~3.5.0",
     "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.4.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-firefox-launcher": "^1.3.0",
     "karma-ie-launcher": "^1.0.0",
-    "karma-jasmine": "~2.0.1",
+    "karma-jasmine": "~3.3.0",
     "karma-sauce-launcher": "^2.0.2",
     "marked": "^0.7.0",
-    "ng-packagr": "~9.0.0",
+    "ng-packagr": "^10.0.0",
     "ngx-build-plus": "^9.0.2",
     "nyc": "14.1.1",
     "prismjs": "1.16.0",
-    "protractor": "~5.4.3",
+    "protractor": "~7.0.0",
     "rimraf": "^3.0.2",
-    "rxjs": "~6.5.4",
+    "rxjs": "~6.5.5",
     "ts-loader": "^6.0.1",
     "ts-node": "~8.3.0",
     "ts-node-dev": "^1.0.0-pre.42",
-    "tslib": "^1.10.0",
-    "tslint": "~5.18.0",
+    "tslint": "~6.1.0",
     "tslint-jasmine-rules": "^1.3.2",
-    "typescript": "~3.7.5",
+    "typescript": "~3.9.5",
     "webpack": "^4.39.3",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.10.3"
   }
 }
