@@ -74,4 +74,4 @@ export const routes: Routes = [
   },
 ];
 
-export const routing: ModuleWithProviders = RouterModule.forRoot(routes, {useHash: true});
+export const routing: ModuleWithProviders<RouterModule> = RouterModule.forRoot(routes, {useHash: true});
